Rename controller secrets -> pullsecret.
