+++ pkg/controller/add_pullsecret.go
@@ -17,10 +17,10 @@ limitations under the License.
 package controller
 
 import (
-	"github.com/Ridecell/ridecell-operator/pkg/controller/secrets"
+	"github.com/Ridecell/ridecell-operator/pkg/controller/pullsecret"
 )
 
 func init() {
 	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
-	AddToManagerFuncs = append(AddToManagerFuncs, secrets.Add)
+	AddToManagerFuncs = append(AddToManagerFuncs, pullsecret.Add)
 }
