+++ pkg/controller/pullsecret/components/defaults.go
