+++ pkg/controller/pullsecret/components/pull_secret.go
