+++ pkg/controller/pullsecret/controller_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package secrets_test
+package pullsecret_test
 
 import (
 	"context"
